@@ -3,50 +3,50 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
+    "postinstall": "sed -i '' 's/#import <fishhook\\/fishhook.h>/#import <React\\/fishhook.h>/' ./node_modules/react-native/Libraries/WebSocket/RCTReconnectingWebSocket.m"
   },
   "dependencies": {
-    "@remobile/react-native-marquee": "^1.0.1",
-    "bugsnag-react-native": "^2.2.3",
+    "axios": "^0.18.0",
+    "bugsnag-react-native": "^2.9.3",
     "firebase": "^3.9.0",
-    "mobx": "^3.2.2",
-    "mobx-react": "^4.2.2",
-    "moment": "^2.18.1",
-    "react": "^16.0.0-alpha.12",
-    "react-native": "^0.47.1",
-    "react-native-admob": "^1.3.0",
-    "react-native-animatable": "^1.2.3",
-    "react-native-aws3": "0.0.8",
-    "react-native-camera": "git+https://github.com/lwansbrough/react-native-camera.git",
-    "react-native-camera-roll-picker": "^1.1.9",
-    "react-native-code-push": "^5.2.1",
-    "react-native-fbsdk": "^0.5.0",
-    "react-native-fcm": "^8.0.0",
+    "mobx": "^3.6.2",
+    "mobx-react": "^4.4.3",
+    "moment": "^2.22.1",
+    "prop-types": "^15.6.1",
+    "react": "16.3.1",
+    "react-native": "0.54.2",
+    "react-native-animatable": "^1.2.4",
+    "react-native-camera": "git+https://git@github.com/react-native-community/react-native-camera.git",
+    "react-native-camera-roll-picker": "^1.2.3",
+    "react-native-code-push": "^5.3.2",
+    "react-native-fbsdk": "^0.7.0",
+    "react-native-fcm": "^14.1.4",
     "react-native-fetch-blob": "^0.10.5",
-    "react-native-form-generator": "^0.9.9",
-    "react-native-gifted-chat": "^0.2.9",
-    "react-native-google-signin": "^0.10.0",
+    "react-native-firebase": "^4.0.6",
+    "react-native-gifted-chat": "^0.4.3",
+    "react-native-google-sign-in": "^1.0.16",
+    "react-native-google-signin": "^0.12.0",
     "react-native-image-resizer": "^1.0.0",
-    "react-native-in-app-utils": "^5.2.3",
-    "react-native-instagram-login": "^1.0.5",
-    "react-native-keyboard-aware-scroll-view": "^0.4.4",
+    "react-native-in-app-utils": "^6.0.0",
+    "react-native-keyboard-aware-scroll-view": "^0.5.0",
     "react-native-lightbox": "^0.7.0",
-    "react-native-loading-spinner-overlay": "^0.5.0",
-    "react-native-nav": "^2.0.0",
-    "react-native-picker": "^4.1.2",
-    "react-native-popup-dialog": "^0.9.29",
+    "react-native-loading-spinner-overlay": "^0.5.2",
+    "react-native-nav": "^2.0.2",
+    "react-native-picker": "^4.3.5",
+    "react-native-popup-dialog": "^0.9.41",
     "react-native-progress": "^3.4.0",
     "react-native-progress-bar": "^0.1.2",
-    "react-native-root-toast": "^1.0.5",
+    "react-native-root-toast": "^3.0.1",
     "react-native-scrollable-tab-view": "^0.8.0",
-    "react-native-simple-radio-button": "^2.7.0",
-    "react-native-sound": "^0.10.0",
-    "react-native-star-rating": "^1.0.7",
-    "react-native-swiper": "^1.5.4",
-    "react-native-vector-icons": "^4.0.1",
-    "react-native-video": "^1.0.0",
-    "react-navigation": "^1.5.8"
+    "react-native-simple-radio-button": "^2.7.1",
+    "react-native-sound": "^0.10.9",
+    "react-native-star-rating": "^1.0.9",
+    "react-native-swiper": "^1.5.13",
+    "react-native-vector-icons": "^4.6.0",
+    "react-native-video": "^2.0.0",
+    "react-navigation": "^1.5.11",
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0"
   },
   "devDependencies": {
     "babel-jest": "19.0.0",
