@@ -1,22 +1,12 @@
 #import <Foundation/Foundation.h>
 
-// NS_SWIFT_NAME can only translate factory methods before the iOS 9.3 SDK.
-// Wrap it in our own macro if it's a non-compatible SDK.
-#ifndef FIR_SWIFT_NAME
-#ifdef __IPHONE_9_3
-#define FIR_SWIFT_NAME(X) NS_SWIFT_NAME(X)
-#else
-#define FIR_SWIFT_NAME(X)  // Intentionally blank.
-#endif  // #ifdef __IPHONE_9_3
-#endif  // #ifndef FIR_SWIFT_NAME
-
 /**
  *  @memberof FIRInstanceID
  *
  *  The scope to be used when fetching/deleting a token for Firebase Messaging.
  */
-FOUNDATION_EXPORT NSString * __nonnull const kFIRInstanceIDScopeFirebaseMessaging
-    FIR_SWIFT_NAME(InstanceIDScopeFirebaseMessaging);
+FOUNDATION_EXPORT NSString *__nonnull const kFIRInstanceIDScopeFirebaseMessaging
+    NS_SWIFT_NAME(InstanceIDScopeFirebaseMessaging);
 
 #if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
 /**
@@ -28,7 +18,7 @@ FOUNDATION_EXPORT NSString * __nonnull const kFIRInstanceIDScopeFirebaseMessagin
  *  to control the rate of token updates on application servers.
  */
 FOUNDATION_EXPORT const NSNotificationName __nonnull kFIRInstanceIDTokenRefreshNotification
-    FIR_SWIFT_NAME(InstanceIDTokenRefresh);
+    NS_SWIFT_NAME(InstanceIDTokenRefresh);
 #else
 /**
  *  Called when the system determines that tokens need to be refreshed.
@@ -38,8 +28,8 @@ FOUNDATION_EXPORT const NSNotificationName __nonnull kFIRInstanceIDTokenRefreshN
  *  Instance ID service will throttle the refresh event across all devices
  *  to control the rate of token updates on application servers.
  */
-FOUNDATION_EXPORT NSString * __nonnull const kFIRInstanceIDTokenRefreshNotification
-    FIR_SWIFT_NAME(InstanceIDTokenRefreshNotification);
+FOUNDATION_EXPORT NSString *__nonnull const kFIRInstanceIDTokenRefreshNotification
+    NS_SWIFT_NAME(InstanceIDTokenRefreshNotification);
 #endif  // defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
 
 /**
@@ -54,9 +44,8 @@ FOUNDATION_EXPORT NSString * __nonnull const kFIRInstanceIDTokenRefreshNotificat
  *               failed. See the error codes below for a more detailed
  *               description.
  */
-typedef void(^FIRInstanceIDTokenHandler)( NSString * __nullable token, NSError * __nullable error)
-    FIR_SWIFT_NAME(InstanceIDTokenHandler);
-
+typedef void (^FIRInstanceIDTokenHandler)(NSString *__nullable token, NSError *__nullable error)
+    NS_SWIFT_NAME(InstanceIDTokenHandler);
 
 /**
  *  @related FIRInstanceID
@@ -67,8 +56,8 @@ typedef void(^FIRInstanceIDTokenHandler)( NSString * __nullable token, NSError *
  *  @param error The error describing why deleting the token failed.
  *               See the error codes below for a more detailed description.
  */
-typedef void(^FIRInstanceIDDeleteTokenHandler)(NSError * __nullable error)
-    FIR_SWIFT_NAME(InstanceIDDeleteTokenHandler);
+typedef void (^FIRInstanceIDDeleteTokenHandler)(NSError *__nullable error)
+    NS_SWIFT_NAME(InstanceIDDeleteTokenHandler);
 
 /**
  *  @related FIRInstanceID
@@ -80,8 +69,8 @@ typedef void(^FIRInstanceIDDeleteTokenHandler)(NSError * __nullable error)
  *                  while creating an identity.
  *  @param error    The error if fetching the identity fails else nil.
  */
-typedef void(^FIRInstanceIDHandler)(NSString * __nullable identity, NSError * __nullable error)
-    FIR_SWIFT_NAME(InstanceIDHandler);
+typedef void (^FIRInstanceIDHandler)(NSString *__nullable identity, NSError *__nullable error)
+    NS_SWIFT_NAME(InstanceIDHandler);
 
 /**
  *  @related FIRInstanceID
@@ -92,11 +81,11 @@ typedef void(^FIRInstanceIDHandler)(NSString * __nullable identity, NSError * __
  *  @param error The error if deleting the identity and all the tokens associated with
  *               it fails else nil.
  */
-typedef void(^FIRInstanceIDDeleteHandler)(NSError * __nullable error)
-    FIR_SWIFT_NAME(InstanceIDDeleteHandler);
+typedef void (^FIRInstanceIDDeleteHandler)(NSError *__nullable error)
+    NS_SWIFT_NAME(InstanceIDDeleteHandler);
 
 /**
- * @enum FIRInstanceIDError
+ * Public errors produced by InstanceID.
  */
 typedef NS_ENUM(NSUInteger, FIRInstanceIDError) {
   // Http related errors.
@@ -122,7 +111,7 @@ typedef NS_ENUM(NSUInteger, FIRInstanceIDError) {
 
   /// InvalidRequest -- Some parameters of the request were invalid.
   FIRInstanceIDErrorInvalidRequest = 7,
-} FIR_SWIFT_NAME(InstanceIDError);
+} NS_SWIFT_NAME(InstanceIDError);
 
 /**
  *  The APNS token type for the app. If the token type is set to `UNKNOWN`
@@ -136,12 +125,16 @@ typedef NS_ENUM(NSInteger, FIRInstanceIDAPNSTokenType) {
   FIRInstanceIDAPNSTokenTypeSandbox,
   /// Production token type.
   FIRInstanceIDAPNSTokenTypeProd,
-} FIR_SWIFT_NAME(InstanceIDAPNSTokenType)
+} NS_SWIFT_NAME(InstanceIDAPNSTokenType)
     __deprecated_enum_msg("Use FIRMessaging's APNSToken property instead.");
 
 /**
  *  Instance ID provides a unique identifier for each app instance and a mechanism
  *  to authenticate and authorize actions (for example, sending an FCM message).
+ *
+ *  Once an InstanceID is generated, the library periodically sends information about the
+ *  application and the device where it's running to the Firebase backend. To stop this. see
+ *  `[FIRInstanceID deleteIDWithHandler:]`.
  *
  *  Instance ID is long lived but, may be reset if the device is not used for
  *  a long time or the Instance ID service detects a problem.
@@ -154,7 +147,7 @@ typedef NS_ENUM(NSInteger, FIRInstanceIDAPNSTokenType) {
  *  services associated with the app, call
  *  `[FIRInstanceID tokenWithAuthorizedEntity:scope:options:handler]`.
  */
-FIR_SWIFT_NAME(InstanceID)
+NS_SWIFT_NAME(InstanceID)
 @interface FIRInstanceID : NSObject
 
 /**
@@ -162,7 +155,7 @@ FIR_SWIFT_NAME(InstanceID)
  *
  *  @return A shared instance of FIRInstanceID.
  */
-+ (nonnull instancetype)instanceID FIR_SWIFT_NAME(instanceID());
++ (nonnull instancetype)instanceID NS_SWIFT_NAME(instanceID());
 
 /**
  *  Unavailable. Use +instanceID instead.
@@ -181,15 +174,15 @@ FIR_SWIFT_NAME(InstanceID)
  */
 - (void)setAPNSToken:(nonnull NSData *)token
                 type:(FIRInstanceIDAPNSTokenType)type
-      __deprecated_msg("Use FIRMessaging's APNSToken property instead.");
+    __deprecated_msg("Use FIRMessaging's APNSToken property instead.");
 
 #pragma mark - Tokens
 
 /**
  *  Returns a Firebase Messaging scoped token for the firebase app.
  *
- *  @return Null Returns null if the device has not yet been registerd with
- *          Firebase Message else returns a valid token.
+ *  @return Returns the stored token if the device has registered with Firebase Messaging, otherwise
+ *          returns nil.
  */
 - (nullable NSString *)token;
 
@@ -204,6 +197,9 @@ FIR_SWIFT_NAME(InstanceID)
  *  we invoke the completion callback with nil `token` and the appropriate
  *  error.
  *
+ *  This generates an Instance ID if it does not exist yet, which starts periodically sending
+ *  information to the Firebase backend (see `[FIRInstanceID getIDWithHandler:]`).
+ *
  *  Note, you can only have one `token` or `deleteToken` call for a given
  *  authorizedEntity and scope at any point of time. Making another such call with the
  *  same authorizedEntity and scope before the last one finishes will result in an
@@ -215,11 +211,14 @@ FIR_SWIFT_NAME(InstanceID)
  *  @param scope            Action authorized for authorizedEntity.
  *  @param options          The extra options to be sent with your token request. The
  *                          value for the `apns_token` should be the NSData object
- *                          passed to UIApplication's
+ *                          passed to the UIApplicationDelegate's
  *                          `didRegisterForRemoteNotificationsWithDeviceToken` method.
- *                          All other keys and values in the options dict need to be
- *                          instances of NSString or else they will be discarded. Bundle
- *                          keys starting with 'GCM.' and 'GOOGLE.' are reserved.
+ *                          The value for `apns_sandbox` should be a boolean (or an
+ *                          NSNumber representing a BOOL in Objective C) set to true if
+ *                          your app is a debug build, which means that the APNs
+ *                          device token is for the sandbox environment. It should be
+ *                          set to false otherwise. If the `apns_sandbox` key is not
+ *                          provided, an automatically-detected value shall be used.
  *  @param handler          The callback handler which is invoked when the token is
  *                          successfully fetched. In case of success a valid `token` and
  *                          `nil` error are returned. In case of any error the `token`
@@ -261,19 +260,30 @@ FIR_SWIFT_NAME(InstanceID)
  *  instance. If the identifier has been revoked or has expired, this method will
  *  return a new identifier.
  *
+ *  Once an InstanceID is generated, the library periodically sends information about the
+ *  application and the device where it's running to the Firebase backend. To stop this. see
+ *  `[FIRInstanceID deleteIDWithHandler:]`.
  *
  *  @param handler The handler to invoke once the identifier has been fetched.
  *                 In case of error an appropriate error object is returned else
  *                 a valid identifier is returned and a valid identifier for the
  *                 application instance.
  */
-- (void)getIDWithHandler:(nonnull FIRInstanceIDHandler)handler
-    FIR_SWIFT_NAME(getID(handler:));
+- (void)getIDWithHandler:(nonnull FIRInstanceIDHandler)handler NS_SWIFT_NAME(getID(handler:));
 
 /**
  *  Resets Instance ID and revokes all tokens.
+ *
+ *  This method also triggers a request to fetch a new Instance ID and Firebase Messaging scope
+ *  token. Please listen to kFIRInstanceIDTokenRefreshNotification when the new ID and token are
+ *  ready.
+ *
+ *  This stops the periodic sending of data to the Firebase backend that began when the Instance ID
+ *  was generated. No more data is sent until another library calls Instance ID internally again
+ *  (like FCM, RemoteConfig or Analytics) or user explicitly calls Instance ID APIs to get an
+ *  Instance ID and token again.
  */
 - (void)deleteIDWithHandler:(nonnull FIRInstanceIDDeleteHandler)handler
-    FIR_SWIFT_NAME(deleteID(handler:));
+    NS_SWIFT_NAME(deleteID(handler:));
 
 @end
