@@ -126,7 +126,7 @@ typedef GPB_ENUM(GPBTimestamp_FieldNumber) {
  * to this format using [`strftime`](https://docs.python.org/2/library/time.html#time.strftime)
  * with the time format spec '%Y-%m-%dT%H:%M:%S.%fZ'. Likewise, in Java, one
  * can use the Joda Time's [`ISODateTimeFormat.dateTime()`](
- * http://joda-time.sourceforge.net/apidocs/org/joda/time/format/ISODateTimeFormat.html#dateTime())
+ * http://www.joda.org/joda-time/apidocs/org/joda/time/format/ISODateTimeFormat.html#dateTime--)
  * to obtain a formatter capable of generating timestamps in this format.
  **/
 @interface GPBTimestamp : GPBMessage
