@@ -10,7 +10,8 @@ import {
   ImageBackground,
   Alert
 } from 'react-native';
-
+import PropTypes from 'prop-types';
+import axios from 'axios';
 import firebase from 'firebase';
 import Backend from '../Backend';
 import { NavigationActions } from 'react-navigation'
@@ -118,26 +119,23 @@ export default class Leader extends React.Component {
 
   componentDidMount() {
 
-    fetch('https://eventfluxbot.herokuapp.com/appapi/getLeaders', {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        uid: Backend.getUid(),
-        yeni:true
-      })
+    axios.post('https://eventfluxbot.herokuapp.com/appapi/getLeaders', {
+      uid: Backend.getUid(),
+      yeni:true
     })
-    .then((response) => response.json())
-     .then((responseJson) => {
-        var leaders=Array.from(responseJson.leaders)
-        var weeks=Array.from(responseJson.weeks)
-        var weekResults=Array.from(responseJson.weekresults)
-        this.setState({leaders:leaders,weeks:weeks,weekResults:weekResults});
+    .then( (response) => {
+      var responseJson=response.data
+
+      var leaders=Array.from(responseJson.leaders)
+      var weeks=Array.from(responseJson.weeks)
+      var weekResults=Array.from(responseJson.weekresults)
+      this.setState({leaders:leaders,weeks:weeks,weekResults:weekResults});
+    })
+    .catch(function (error) {
+
+    });
 
 
-     })
 
   }
   componentDidUpdate() {
@@ -262,78 +260,73 @@ export default class Leader extends React.Component {
   showProfPopup = (fireid,profPhoto) =>{
     this.popupDialog.show()
     this.setState({profinfo:null})
-    fetch('https://eventfluxbot.herokuapp.com/webhook/getAppUser', {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        uid: fireid,
-      })
+    axios.post('https://eventfluxbot.herokuapp.com/webhook/getAppUser', {
+      uid: fireid,
     })
-    .then((response) => response.json())
-   .then((responseJson) => {
-     //alert(responseJson.profile_pic)
+    .then( (response) => {
+      var responseJson=response.data
+      responseJson.profile_pic?responseJson.profile_pic={uri:responseJson.profile_pic}:profPhoto?responseJson.profile_pic={uri:profPhoto}:responseJson.gender=="female"?responseJson.profile_pic=require('../static/images/femaleAvatar.png'):responseJson.profile_pic=require('../static/images/maleAvatar.png')
+       var meslek =''
+       switch(responseJson.workStatus) {
+         case 1:
+             meslek='Çalışıyor';
+             break;
+         case 2:
+             meslek='İş arıyor';
+             break;
+         case 3:
+             meslek='Öğrenci';
+             break;
+         case 4:
+             meslek='Çalışmıyor';
+             break;
+         case 5:
+             meslek='Kamuda Çalışıyorum';
+             break;
+         case 6:
+             meslek='Özel Sektör';
+             break;
+         case 7:
+             meslek='Kendi İşim';
+             break;
+       }
+       var iliski =''
+       switch(responseJson.relStatus) {
+           case "0":
+               iliski='İlişkisi Yok';
+               break;
+           case "1":
+               iliski='Sevgilisi Var';
+               break;
+           case "2":
+               iliski='Evli';
+               break;
+           case "3":
+               iliski='Nişanlı';
+               break;
+           case "4":
+               iliski='Platonik';
+               break;
+           case "5":
+               iliski='Ayrı Yaşıyor';
+               break;
+           case "6":
+               iliski='Yeni Ayrılmış';
+               break;
+           case "7":
+               iliski='Boşanmış';
+               break;
 
-    responseJson.profile_pic?responseJson.profile_pic={uri:responseJson.profile_pic}:profPhoto?responseJson.profile_pic={uri:profPhoto}:responseJson.gender=="female"?responseJson.profile_pic=require('../static/images/femaleAvatar.png'):responseJson.profile_pic=require('../static/images/maleAvatar.png')
-     var meslek =''
-     switch(responseJson.workStatus) {
-       case 1:
-           meslek='Çalışıyor';
-           break;
-       case 2:
-           meslek='İş arıyor';
-           break;
-       case 3:
-           meslek='Öğrenci';
-           break;
-       case 4:
-           meslek='Çalışmıyor';
-           break;
-       case 5:
-           meslek='Kamuda Çalışıyorum';
-           break;
-       case 6:
-           meslek='Özel Sektör';
-           break;
-       case 7:
-           meslek='Kendi İşim';
-           break;
-     }
-     var iliski =''
-     switch(responseJson.relStatus) {
-         case "0":
-             iliski='İlişkisi Yok';
-             break;
-         case "1":
-             iliski='Sevgilisi Var';
-             break;
-         case "2":
-             iliski='Evli';
-             break;
-         case "3":
-             iliski='Nişanlı';
-             break;
-         case "4":
-             iliski='Platonik';
-             break;
-         case "5":
-             iliski='Ayrı Yaşıyor';
-             break;
-         case "6":
-             iliski='Yeni Ayrılmış';
-             break;
-         case "7":
-             iliski='Boşanmış';
-             break;
+       }
+       responseJson.iliski=iliski
+       responseJson.meslek=meslek
+          this.setState({profinfo:responseJson});
 
-     }
-     responseJson.iliski=iliski
-     responseJson.meslek=meslek
-        this.setState({profinfo:responseJson});
+    })
+    .catch(function (error) {
 
-      })
+    });
+
   }
 
   renderLeaders = () => {
